dockershim: clean up the legacy interface
